# Conflicts:
#	app/code/core/Mage/CatalogRule/Model/Resource/Rule.php

+++ app/code/core/Mage/Admin/Model/Resource/User.php
@@ -1,26 +1,22 @@
 <?php
+
 /**
- * OpenMage
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available at https://opensource.org/license/osl-3-0-php
- *
- * @category   Mage
+ * @copyright  For copyright and license information, read the COPYING.txt file.
+ * @link       /COPYING.txt
+ * @license    Open Software License (OSL 3.0)
  * @package    Mage_Admin
- * @copyright  Copyright (c) 2006-2020 Magento, Inc. (https://www.magento.com)
- * @copyright  Copyright (c) 2017-2023 The OpenMage Contributors (https://www.openmage.org)
- * @license    https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
 /**
  * ACL user resource
  *
- * @category   Mage
  * @package    Mage_Admin
  */
 class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstract
 {
+    /**
+     * @inheritDoc
+     */
     protected function _construct()
     {
         $this->_init('admin/user', 'user_id');
@@ -29,16 +25,17 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * @return $this
      */
+    #[Override]
     protected function _initUniqueFields()
     {
         $this->_uniqueFields = [
             [
                 'field' => 'email',
-                'title' => Mage::helper('adminhtml')->__('Email')
+                'title' => Mage::helper('adminhtml')->__('Email'),
             ],
             [
                 'field' => 'username',
-                'title' => Mage::helper('adminhtml')->__('User Name')
+                'title' => Mage::helper('adminhtml')->__('User Name'),
             ],
         ];
         return $this;
@@ -47,8 +44,8 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Authenticate user by $username and $password
      *
-     * @param Mage_Admin_Model_User $user
      * @return $this
+     * @throws Zend_Db_Adapter_Exception
      */
     public function recordLogin(Mage_Admin_Model_User $user)
     {
@@ -56,7 +53,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
 
         $data = [
             'logdate' => Varien_Date::now(),
-            'lognum'  => $user->getLognum() + 1
+            'lognum'  => $user->getLognum() + 1,
         ];
 
         $condition = [
@@ -71,8 +68,8 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Load data by specified username
      *
-     * @param string $username
-     * @return false|array
+     * @param  string      $username
+     * @return array|false
      */
     public function loadByUsername($username)
     {
@@ -83,7 +80,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             ->where('username=:username');
 
         $binds = [
-            'username' => $username
+            'username' => $username,
         ];
 
         return $adapter->fetchRow($select, $binds);
@@ -92,7 +89,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if user is assigned to any role
      *
-     * @param int|Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  int|Mage_Admin_Model_User|Mage_Core_Model_Abstract|string $user
      * @return null|array
      */
     public function hasAssigned2Role($user)
@@ -119,100 +116,99 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             ];
 
             return $adapter->fetchAll($select, $binds);
-        } else {
-            return null;
         }
-    }
 
-    /**
-     * Encrypt password
-     *
-     * @param string $pwStr
-     * @return string
-     */
-    private function _encryptPassword($pwStr)
-    {
-        return Mage::helper('core')->getHash($pwStr, Mage_Admin_Model_User::HASH_SALT_LENGTH);
+        return null;
     }
 
     /**
      * Set created/modified values before user save
      *
-     * @param Mage_Admin_Model_User $user
+     * @param Mage_Admin_Model_User $object
      * @inheritDoc
      */
-    protected function _beforeSave(Mage_Core_Model_Abstract $user)
+    #[Override]
+    protected function _beforeSave(Mage_Core_Model_Abstract $object)
     {
-        if ($user->isObjectNew()) {
-            $user->setCreated($this->formatDate(true));
+        if ($object->isObjectNew()) {
+            $object->setCreated($this->formatDate(true));
         }
-        $user->setModified($this->formatDate(true));
 
-        return parent::_beforeSave($user);
+        $object->setModified($this->formatDate(true));
+
+        return parent::_beforeSave($object);
     }
 
     /**
      * Unserialize user extra data after user save
      *
-     * @param Mage_Core_Model_Abstract $user
+     * @param  Mage_Admin_Model_User $object
      * @return $this
      */
-    protected function _afterSave(Mage_Core_Model_Abstract $user)
+    #[Override]
+    protected function _afterSave(Mage_Core_Model_Abstract $object)
     {
-        $this->_unserializeExtraData($user);
+        $this->_unserializeExtraData($object);
         return $this;
     }
 
     /**
      * Unserialize user extra data after user load
      *
+     * @param Mage_Admin_Model_User $object
      * @inheritDoc
      */
-    protected function _afterLoad(Mage_Core_Model_Abstract $user)
+    #[Override]
+    protected function _afterLoad(Mage_Core_Model_Abstract $object)
     {
-        return parent::_afterLoad($this->_unserializeExtraData($user));
+        return parent::_afterLoad($this->_unserializeExtraData($object));
     }
 
     /**
      * Delete user role record with user
      *
-     * @param Mage_Core_Model_Abstract $user
+     * @param  Mage_Admin_Model_User $object
      * @return $this
      * @throws Exception
+     * @throws Throwable
      */
-    public function delete(Mage_Core_Model_Abstract $user)
+    #[Override]
+    public function delete(Mage_Core_Model_Abstract $object)
     {
-        $this->_beforeDelete($user);
+        $this->_beforeDelete($object);
         $adapter = $this->_getWriteAdapter();
 
-        $uid = $user->getId();
+        $uid = $object->getId();
         $adapter->beginTransaction();
         try {
             $conditions = [
-                'user_id = ?' => $uid
+                'user_id = ?' => $uid,
             ];
 
             $adapter->delete($this->getMainTable(), $conditions);
             $adapter->delete($this->getTable('admin/role'), $conditions);
             $adapter->commit();
-        } catch (Throwable $e) {
+        } catch (Throwable $throwable) {
             $adapter->rollBack();
-            throw $e;
+            throw $throwable;
         }
-        $this->_afterDelete($user);
+
+        $this->_afterDelete($object);
         return $this;
     }
 
     /**
      * TODO: unify _saveRelations() and add() methods, they make same things
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
-     * @return $this|Mage_Core_Model_Abstract
+     * @param  Mage_Admin_Model_User                                $user
+     * @return $this|Mage_Admin_Model_User|Mage_Core_Model_Abstract
+     * @throws Mage_Core_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function _saveRelations(Mage_Core_Model_Abstract $user)
     {
         $rolesIds = $user->getRoleIds();
-        if (!is_array($rolesIds) || count($rolesIds) == 0) {
+        if (!is_array($rolesIds) || $rolesIds === []) {
             return $user;
         }
 
@@ -227,11 +223,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $adapter->delete($this->getTable('admin/role'), $conditions);
             foreach ($rolesIds as $rid) {
                 $rid = (int) $rid;
-                if ($rid > 0) {
-                    $role = Mage::getModel('admin/role')->load($rid);
-                } else {
-                    $role = new Varien_Object(['tree_level' => 0]);
-                }
+                $role = $rid > 0 ? Mage::getModel('admin/role')->load($rid) : new Varien_Object(['tree_level' => 0]);
 
                 $data = new Varien_Object([
                     'parent_id'  => $rid,
@@ -239,7 +231,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                     'sort_order' => 0,
                     'role_type'  => Mage_Admin_Model_Acl::ROLE_TYPE_USER,
                     'user_id'    => $user->getId(),
-                    'role_name'  => $user->getFirstname()
+                    'role_name'  => $user->getFirstname(),
                 ]);
 
                 $insertData = $this->_prepareDataForTable($data, $this->getTable('admin/role'));
@@ -250,13 +242,14 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                 // reload acl on next user http request
                 $this->saveReloadAclFlag($user, 1);
             }
+
             $adapter->commit();
-        } catch (Mage_Core_Exception $e) {
+        } catch (Mage_Core_Exception $mageCoreException) {
             $adapter->rollBack();
-            throw $e;
-        } catch (Exception $e) {
+            throw $mageCoreException;
+        } catch (Exception $exception) {
             $adapter->rollBack();
-            throw $e;
+            throw $exception;
         }
 
         return $this;
@@ -265,7 +258,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Get user roles
      *
-     * @param Mage_Core_Model_Abstract $user
+     * @param  Mage_Admin_Model_User $user
      * @return array
      */
     public function getRoles(Mage_Core_Model_Abstract $user)
@@ -281,7 +274,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                     ->joinLeft(
                         ['ar' => $table],
                         "(ar.role_id = {$table}.parent_id and ar.role_type = 'G')",
-                        ['role_id']
+                        ['role_id'],
                     )
                     ->where("{$table}.user_id = :user_id");
 
@@ -301,18 +294,20 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Save user roles
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  Mage_Admin_Model_User     $user
      * @return $this
+     * @throws Zend_Cache_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function add(Mage_Core_Model_Abstract $user)
     {
         $dbh = $this->_getWriteAdapter();
         $aRoles = $this->hasAssigned2Role($user);
-        if (count($aRoles)) {
-            foreach ($aRoles as $idx => $data) {
+        if ($aRoles && count($aRoles)) {
+            foreach ($aRoles as $data) {
                 $dbh->delete(
                     $this->getTable('admin/role'),
-                    ['role_id = ?' => $data['role_id']]
+                    ['role_id = ?' => $data['role_id']],
                 );
             }
         }
@@ -329,7 +324,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             'sort_order' => 0,
             'role_type'  => Mage_Admin_Model_Acl::ROLE_TYPE_USER,
             'user_id'    => $user->getUserId(),
-            'role_name'  => $user->getFirstname()
+            'role_name'  => $user->getFirstname(),
         ]);
 
         $insertData = $this->_prepareDataForTable($data, $this->getTable('admin/role'));
@@ -346,7 +341,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Delete user role
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  Mage_Admin_Model_User $user
      * @return $this
      */
     public function deleteFromRole(Mage_Core_Model_Abstract $user)
@@ -354,6 +349,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         if ($user->getUserId() <= 0) {
             return $this;
         }
+
         if ($user->getRoleId() <= 0) {
             return $this;
         }
@@ -372,7 +368,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if role user exists
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  Mage_Admin_Model_User $user
      * @return array
      */
     public function roleUserExists(Mage_Core_Model_Abstract $user)
@@ -392,15 +388,15 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                 ->where('user_id = :user_id');
 
             return $dbh->fetchCol($select, $binds);
-        } else {
-            return [];
         }
+
+        return [];
     }
 
     /**
      * Check if user exists
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  Mage_Admin_Model_User $user
      * @return array|false
      */
     public function userExists(Mage_Core_Model_Abstract $user)
@@ -424,9 +420,10 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Save user extra data
      *
-     * @param Mage_Core_Model_Abstract $object
-     * @param string $data
+     * @param  Mage_Admin_Model_User     $object
+     * @param  string                    $data
      * @return $this
+     * @throws Zend_Db_Adapter_Exception
      */
     public function saveExtra($object, $data)
     {
@@ -434,7 +431,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $this->_getWriteAdapter()->update(
                 $this->getMainTable(),
                 ['extra' => $data],
-                ['user_id = ?' => (int) $object->getId()]
+                ['user_id = ?' => (int) $object->getId()],
             );
         }
 
@@ -444,9 +441,11 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Set reload ACL flag
      *
-     * @param Mage_Core_Model_Abstract $object
-     * @param int $flag
+     * @param  Mage_Admin_Model_User     $object
+     * @param  int                       $flag
      * @return $this
+     * @throws Zend_Cache_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function saveReloadAclFlag($object, $flag)
     {
@@ -454,13 +453,13 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $this->_getWriteAdapter()->update(
                 $this->getMainTable(),
                 ['reload_acl_flag' => $flag],
-                ['user_id = ?' => (int) $object->getId()]
+                ['user_id = ?' => (int) $object->getId()],
             );
             if ($flag) {
                 // refresh cache menu
                 Mage::app()->getCache()->clean(
                     Zend_Cache::CLEANING_MODE_MATCHING_TAG,
-                    [Mage_Adminhtml_Block_Page_Menu::CACHE_TAGS]
+                    [Mage_Adminhtml_Block_Page_Menu::CACHE_TAGS],
                 );
             }
         }
@@ -471,7 +470,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Unserializes user extra data
      *
-     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  Mage_Admin_Model_User    $user
      * @return Mage_Core_Model_Abstract
      */
     protected function _unserializeExtraData(Mage_Core_Model_Abstract $user)
@@ -479,9 +478,10 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         try {
             $unsterilizedData = Mage::helper('core/unserializeArray')->unserialize($user->getExtra());
             $user->setExtra($unsterilizedData);
-        } catch (Exception $e) {
+        } catch (Exception) {
             $user->setExtra(false);
         }
+
         return $user;
     }
 }
