+++ app/code/core/Mage/Captcha/Model/Observer.php
@@ -1,22 +1,17 @@
 <?php
+
 /**
- * OpenMage
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available at https://opensource.org/license/osl-3-0-php
- *
- * @category   Mage
+ * @copyright  For copyright and license information, read the COPYING.txt file.
+ * @link       /COPYING.txt
+ * @license    Open Software License (OSL 3.0)
  * @package    Mage_Captcha
- * @copyright  Copyright (c) 2006-2020 Magento, Inc. (https://www.magento.com)
- * @copyright  Copyright (c) 2019-2023 The OpenMage Contributors (https://www.openmage.org)
- * @license    https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
+
+use Carbon\Carbon;
 
 /**
  * Captcha Observer
  *
- * @category   Mage
  * @package    Mage_Captcha
  */
 class Mage_Captcha_Model_Observer
@@ -24,7 +19,7 @@ class Mage_Captcha_Model_Observer
     /**
      * Check Captcha On Forgot Password Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkForgotpassword($observer)
@@ -39,13 +34,14 @@ class Mage_Captcha_Model_Observer
                 $controller->getResponse()->setRedirect(Mage::getUrl('*/*/forgotpassword'));
             }
         }
+
         return $this;
     }
 
     /**
      * Check Captcha On User Login Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkUserLogin($observer)
@@ -66,6 +62,7 @@ class Mage_Captcha_Model_Observer
                 $controller->getResponse()->setRedirect($url);
             }
         }
+
         $captchaModel->logAttempt($login);
         return $this;
     }
@@ -73,7 +70,6 @@ class Mage_Captcha_Model_Observer
     /**
      * Check Captcha On Register User Page
      *
-     * @param Varien_Event_Observer $observer
      * @return $this
      */
     public function checkUserCreate(Varien_Event_Observer $observer)
@@ -89,13 +85,13 @@ class Mage_Captcha_Model_Observer
                 $controller->getResponse()->setRedirect(Mage::getUrl('*/*/create'));
             }
         }
+
         return $this;
     }
 
     /**
      * Check Captcha On Checkout as Guest Page
      *
-     * @param Varien_Event_Observer $observer
      * @return $this
      */
     public function checkGuestCheckout(Varien_Event_Observer $observer)
@@ -103,23 +99,22 @@ class Mage_Captcha_Model_Observer
         $formId = 'guest_checkout';
         $captchaModel = Mage::helper('captcha')->getCaptcha($formId);
         $checkoutMethod = Mage::getSingleton('checkout/type_onepage')->getQuote()->getCheckoutMethod();
-        if ($checkoutMethod == Mage_Checkout_Model_Type_Onepage::METHOD_GUEST) {
-            if ($captchaModel->isRequired()) {
-                $controller = $observer->getControllerAction();
-                if (!$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))) {
-                    $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
-                    $result = ['error' => 1, 'message' => Mage::helper('captcha')->__('Incorrect CAPTCHA.')];
-                    $controller->getResponse()->setBody(Mage::helper('core')->jsonEncode($result));
-                }
+        if ($checkoutMethod == Mage_Checkout_Model_Type_Onepage::METHOD_GUEST && $captchaModel->isRequired()) {
+            $controller = $observer->getControllerAction();
+            if (!$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))) {
+                $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
+                $result = ['error' => 1, 'message' => Mage::helper('captcha')->__('Incorrect CAPTCHA.')];
+                $controller->getResponse()->setBody(Mage::helper('core')->jsonEncode($result));
             }
         }
+
         return $this;
     }
 
     /**
      * Check Captcha On Checkout Register Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkRegisterCheckout($observer)
@@ -127,24 +122,24 @@ class Mage_Captcha_Model_Observer
         $formId = 'register_during_checkout';
         $captchaModel = Mage::helper('captcha')->getCaptcha($formId);
         $checkoutMethod = Mage::getSingleton('checkout/type_onepage')->getQuote()->getCheckoutMethod();
-        if ($checkoutMethod == Mage_Checkout_Model_Type_Onepage::METHOD_REGISTER) {
-            if ($captchaModel->isRequired()) {
-                $controller = $observer->getControllerAction();
-                if (!$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))) {
-                    $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
-                    $result = ['error' => 1, 'message' => Mage::helper('captcha')->__('Incorrect CAPTCHA.')];
-                    $controller->getResponse()->setBody(Mage::helper('core')->jsonEncode($result));
-                }
+        if ($checkoutMethod == Mage_Checkout_Model_Type_Onepage::METHOD_REGISTER && $captchaModel->isRequired()) {
+            $controller = $observer->getControllerAction();
+            if (!$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))) {
+                $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
+                $result = ['error' => 1, 'message' => Mage::helper('captcha')->__('Incorrect CAPTCHA.')];
+                $controller->getResponse()->setBody(Mage::helper('core')->jsonEncode($result));
             }
         }
+
         return $this;
     }
 
     /**
      * Check Captcha On User Login Backend Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
+     * @throws Mage_Core_Exception
      */
     public function checkUserLoginBackend($observer)
     {
@@ -152,12 +147,13 @@ class Mage_Captcha_Model_Observer
         $captchaModel = Mage::helper('captcha')->getCaptcha($formId);
         $loginParams = Mage::app()->getRequest()->getPost('login', []);
         $login = $loginParams['username'] ?? null;
-        if ($captchaModel->isRequired($login)) {
-            if (!$captchaModel->isCorrect($this->_getCaptchaString(Mage::app()->getRequest(), $formId))) {
-                $captchaModel->logAttempt($login);
-                Mage::throwException(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
-            }
+        if ($captchaModel->isRequired($login)
+            && !$captchaModel->isCorrect($this->_getCaptchaString(Mage::app()->getRequest(), $formId))
+        ) {
+            $captchaModel->logAttempt($login);
+            Mage::throwException(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
         }
+
         $captchaModel->logAttempt($login);
         return $this;
     }
@@ -175,7 +171,7 @@ class Mage_Captcha_Model_Observer
     /**
      * Check Captcha On User Login Backend Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkUserForgotPasswordBackend($observer)
@@ -186,23 +182,24 @@ class Mage_Captcha_Model_Observer
         $email = (string) $observer->getControllerAction()->getRequest()->getParam('email');
         $params = $observer->getControllerAction()->getRequest()->getParams();
 
-        if (!empty($email) && !empty($params)) {
-            if ($captchaModel->isRequired()) {
-                if (!$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))) {
-                    $this->_getBackendSession()->setEmail((string) $controller->getRequest()->getPost('email'));
-                    $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
-                    $this->_getBackendSession()->addError(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
-                    $controller->getResponse()->setRedirect(Mage::getUrl('*/*/forgotpassword'));
-                }
-            }
+        if (!empty($email)
+            && !empty($params)
+            && $captchaModel->isRequired()
+            && !$captchaModel->isCorrect($this->_getCaptchaString($controller->getRequest(), $formId))
+        ) {
+            $this->_getBackendSession()->setEmail((string) $controller->getRequest()->getPost('email'));
+            $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
+            $this->_getBackendSession()->addError(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
+            $controller->getResponse()->setRedirect(Mage::getUrl('*/*/forgotpassword'));
         }
+
         return $this;
     }
 
     /**
      * Reset Attempts For Frontend
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function resetAttemptForFrontend($observer)
@@ -213,7 +210,7 @@ class Mage_Captcha_Model_Observer
     /**
      * Reset Attempts For Backend
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function resetAttemptForBackend($observer)
@@ -236,27 +233,28 @@ class Mage_Captcha_Model_Observer
      * Delete Expired Captcha Images
      *
      * @return $this
+     * @throws Mage_Core_Exception
      */
     public function deleteExpiredImages()
     {
         foreach (Mage::app()->getWebsites(true) as $website) {
-            $expire = time() - Mage::helper('captcha')->getConfigNode('timeout', $website->getDefaultStore()) * 60;
+            $timeout = (int) Mage::helper('captcha')->getConfigNode('timeout', $website->getDefaultStore());
+            $expire = Carbon::now()->subMinutes($timeout)->getTimestamp();
             $imageDirectory = Mage::helper('captcha')->getImgDir($website);
             foreach (new DirectoryIterator($imageDirectory) as $file) {
-                if ($file->isFile() && pathinfo($file->getFilename(), PATHINFO_EXTENSION) == 'png') {
-                    if ($file->getMTime() < $expire) {
-                        unlink($file->getPathname());
-                    }
+                if ($file->isFile() && pathinfo($file->getFilename(), PATHINFO_EXTENSION) === 'png' && $file->getMTime() < $expire) {
+                    unlink($file->getPathname());
                 }
             }
         }
+
         return $this;
     }
 
     /**
      * Reset Attempts
      *
-     * @param string $login
+     * @param  string $login
      * @return $this
      */
     protected function _resetAttempt($login)
@@ -268,20 +266,20 @@ class Mage_Captcha_Model_Observer
     /**
      * Get Captcha String
      *
-     * @param Mage_Core_Controller_Request_Http $request
-     * @param string $formId
+     * @param  Mage_Core_Controller_Request_Http $request
+     * @param  string                            $formId
      * @return string
      */
     protected function _getCaptchaString($request, $formId)
     {
         $captchaParams = $request->getPost(Mage_Captcha_Helper_Data::INPUT_NAME_FIELD_VALUE);
-        return $captchaParams[$formId];
+        return $captchaParams[$formId] ?? '';
     }
 
     /**
      * Check Captcha On Share Wishlist Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkWishlistSharing($observer)
@@ -295,18 +293,19 @@ class Mage_Captcha_Model_Observer
                 Mage::getSingleton('wishlist/session')->addError(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
                 $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
                 Mage::getSingleton('wishlist/session')->setSharingForm($request->getPost());
-                $wishlistId = (int)$request->getParam('wishlist_id');
+                $wishlistId = (int) $request->getParam('wishlist_id');
                 $controller->getResponse()
                     ->setRedirect(Mage::getUrl('wishlist/index/share/wishlist_id/' . $wishlistId));
             }
         }
+
         return $this;
     }
 
     /**
      * Check Captcha On Email Product To A Friend Page
      *
-     * @param Varien_Event_Observer $observer
+     * @param  Varien_Event_Observer $observer
      * @return $this
      */
     public function checkSendfriendSend($observer)
@@ -320,14 +319,16 @@ class Mage_Captcha_Model_Observer
                 Mage::getSingleton('catalog/session')->addError(Mage::helper('captcha')->__('Incorrect CAPTCHA.'));
                 $controller->setFlag('', Mage_Core_Controller_Varien_Action::FLAG_NO_DISPATCH, true);
                 Mage::getSingleton('catalog/session')->setFormData($request->getPost());
-                $id = (int)$request->getParam('id');
+                $sendId = (int) $request->getParam('id');
                 $catId = $request->getParam('cat_id');
                 if ($catId !== null) {
-                    $id .= '/cat_id/' . (int)$catId;
+                    $sendId .= '/cat_id/' . (int) $catId;
                 }
-                $controller->getResponse()->setRedirect(Mage::getUrl('*/*/send/id/' . $id));
+
+                $controller->getResponse()->setRedirect(Mage::getUrl('*/*/send/id/' . $sendId));
             }
         }
+
         return $this;
     }
 }
